requirements.txt
customtkinter
Pillow
python-dotenv
google-generativeai
